// fe/src/lib/api/generated/model/routineFormRevisionSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface RoutineFormRevisionSummary {
  id: string;
  internal_schema_version: number;
  official_code: string | null;
  official_revision: string | null;
}