// fe/src/lib/api/generated/model/routineInventoryContext.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { RoutineAcademicYearSummary } from './routineAcademicYearSummary';

export interface RoutineInventoryContext {
  academic_year: RoutineAcademicYearSummary;
  course: string;
  full_name: string;
  id: string;
  major: string;
}